--- a/ChangeLog.txt
+++ b/ChangeLog.txt
@@ -1,0 +1,8 @@
+Replaced the StringBuilder code with simple concatination for speed
+Added configurable significant digits
+Moved toolbar code into same DLL (saves 7 k, and no need to multiple DLLs)
+Added simple settings page which can disable the toolbar button in a save
+
+Needed
+
+AVC .version file for version control