--- a/SpeedIndicator/Settings.cs
+++ b/SpeedIndicator/Settings.cs
@@ -1,0 +1,47 @@
+﻿using System;
+using System.IO;
+using System.Collections;
+using System.Collections.Generic;
+using System.Linq;
+using System.Reflection;
+using System.Text;
+using UnityEngine;
+
+
+namespace SpeedUnitChangerToolbar
+{
+    public class SUC : GameParameters.CustomParameterNode
+    {
+        public override string Title { get { return ""; } }
+        public override GameParameters.GameMode GameMode { get { return GameParameters.GameMode.ANY; } }
+        public override string Section { get { return "Patch Manager"; } }
+        public override string DisplaySection { get { return "Patch Manager"; } }
+        public override int SectionOrder { get { return 2; } }
+        public override bool HasPresets { get { return true; } }
+
+
+        [GameParameters.CustomParameterUI("Mod Enabled")]
+        public bool EnabledForSave = true;      // is enabled for this save file
+
+
+        public override void SetDifficultyPreset(GameParameters.Preset preset)
+        {
+        }
+
+        public override bool Enabled(MemberInfo member, GameParameters parameters)
+        {
+            return true;
+        }
+
+        public override bool Interactible(MemberInfo member, GameParameters parameters)
+        {
+            return true;
+        }
+
+        public override IList ValidValues(MemberInfo member)
+        {
+            return null;
+        }
+    }
+
+}