… SpeedUnitChangerToolbar/Properties/AssemblyInfo.cs
… SpeedUnitChangerToolbar/SpeedUnitChangerToolbar.csproj