Replaced the StringBuilder code with simple concatenation for speed.  The old code is still there, but commented out.  See below for explanation

Added configurable significant digits
Moved toolbar code into same DLL (saves 7 k, and no need to multiple DLLs)
Added simple settings page which can disable the toolbar button in a save

Explanation
I recently was running some tests on various methods of string concatenation with regards to speed.  
While StringBuilder is definitely the cleanest in terms of readability, it has a significant performance impact, especially when used in conjunction with String.Format.
I found that the StringBuilder is the slowest of the methods, and since this gets called at least 20x a second, it actually was causing problems for me on my big install, where I have about 220 mods installed.  In some places you actually were using both String.Format AND StringBuilder, which was adding a lot of overhead.
See the following for the results of my testings:

Beginning timing tests
Timing results: Operator + Time: 551 ms/1_000_000 calls 

Timing results: String.Format Time: 995 ms/1_000_000 calls 

Timing results: StringBuilder Time: 653 ms/1_000_000 calls 

Timing results: StringBuilder 2 Time: 400 ms/1_000_000 calls 

Completed timing tests