@linuxgurugamer 2018-01-10
Note, that last timing test I listed was for a very different test, it should be ignored